+++ public/react-services/wz-authentication.ts
@@ -14,10 +14,10 @@ import { WzRequest } from './wz-request';
 import { AppState } from './app-state';
 import jwtDecode from 'jwt-decode';
 import store from '../redux/store';
-import { updateUserPermissions, updateUserRoles, updateWithUserLogged } from '../redux/actions/appStateActions';
+import { updateUserPermissions, updateUserRoles, updateWithUserLogged, updateAllowedAgents } from '../redux/actions/appStateActions';
 import { WAZUH_ROLE_ADMINISTRATOR_ID, WAZUH_ROLE_ADMINISTRATOR_NAME } from '../../common/constants';
 import { getToasts } from '../kibana-services';
-
+import { getAuthorizedAgents } from '../react-services/wz-agents';
 
 export class WzAuthentication{
   private static async login(force=false){
@@ -48,6 +48,10 @@ export class WzAuthentication{
 
       // Decode token and get expiration time
       const jwtPayload = jwtDecode(token);
+
+      //Get allowed agents for the current user
+      const allowedAgents = await getAuthorizedAgents();
+      store.dispatch(updateAllowedAgents(allowedAgents));
 
       // Get user Policies
       const userPolicies = await WzAuthentication.getUserPolicies();